refactor: move table printing out of pkg
